Rename rebuild-dependencies.sh; dont build same version twice

+++ rebuild-dependencies.sh
@@ -74,6 +74,15 @@ else
     bzr branch ${project_repo} ${code_dir}
 fi
 
+existing_revision=$(cat ${dependencies_dir}/project-revision.txt)
+latest_revision=$(bzr-revision-id ${code_dir})
+
+# Make sure this is a new revision
+if [[ "${existing_revision}" == "${latest_revision}" ]]; then
+    echo "New version (${latest_revision}) is the same as the existing version (${existing_revision}). Aborting."
+    exit 1
+fi
+
 # Clear out existing dependencies, to create from scratch again
 rm -r ${dependencies_dir}/*
 
@@ -81,7 +90,7 @@ rm -r ${dependencies_dir}/*
 pip install --exists-action=w --download ${dependencies_dir} -r ${code_dir}/requirements/standard.txt
 
 # Get latest revision number of the project, store it alongside dependencies
-echo $(bzr-revision-id ${code_dir}) > ${dependencies_dir}/project-revision.txt
+echo ${latest_revision} > ${dependencies_dir}/project-revision.txt
 
 # Commit and push all new files
 bzr add ${dependencies_dir}/.
